Rename Context to CommandContext
